+++ src/io/XDRIO.jl
@@ -5,7 +5,7 @@ struct XDRIO{T<:IO} <: RDAIO
     sub::T             # underlying IO stream
     buf::Vector{UInt8} # buffer for strings
 
-    XDRIO(io::T) where {T <: IO} = new{T}(io, Vector{UInt8}(1024))
+    XDRIO(io::T) where {T <: IO} = new{T}(io, Vector{UInt8}(undef, 1024))
 end
 
 readint32(io::XDRIO) = ntoh(read(io.sub, Int32))
@@ -14,7 +14,7 @@ readfloat64(io::XDRIO) = ntoh(read(io.sub, Float64))
 
 readintorNA(io::XDRIO) = readint32(io)
 function readintorNA(io::XDRIO, n::RVecLength)
-    v = read(io.sub, Int32, n)
+    v = read!(io.sub, Vector{Int32}(undef, n))
     map!(ntoh, v, v)
 end
 
@@ -22,7 +22,7 @@ end
 # R's NA is silently converted to NaN when the value is loaded in the register(?)
 #readfloatorNA(io::XDRIO) = readfloat64(io)
 function readfloatorNA(io::XDRIO, n::RVecLength)
-    v = read(io.sub, Float64, n)
+    v = read!(io.sub, Vector{Float64}(undef, n))
     map!(ntoh, v, v)
 end